+++ src/modules/schedules/schedules.module.ts
@@ -6,6 +6,7 @@ import { Schedule } from 'src/entities/schedule.entity';
 import { HttpModule } from '@nestjs/axios';
 import { ConfigModule } from '@nestjs/config';
 import { CategoriesModule } from '../categories/categories.module';
+import { VoiceTranscriptionService } from './voice-transcription.service';
 @Module({
   imports: [
     TypeOrmModule.forFeature([Schedule]),
@@ -14,6 +15,6 @@ import { CategoriesModule } from '../categories/categories.module';
     CategoriesModule,
   ],
   controllers: [SchedulesController],
-  providers: [SchedulesService],
+  providers: [SchedulesService, VoiceTranscriptionService],
 })
 export class SchedulesModule {}
