+++ src/modules/schedules/schedules.service.ts
@@ -14,13 +14,11 @@ import { MonthQueryDto } from './dto/month-query-schedule.dto';
 import { WeekQueryDto } from './dto/week-query-schedule.dto';
 import { plainToInstance } from 'class-transformer';
 import { HttpService } from '@nestjs/axios';
-import { lastValueFrom } from 'rxjs';
-import FormData from 'form-data';
 import { ConfigService } from '@nestjs/config';
 import OpenAI from 'openai';
 import { Category } from '@/entities/category.entity';
-import * as jwt from 'jsonwebtoken';
 import { VoiceScheduleResponseDto } from './dto/voice-schedule.dto';
+import { VoiceTranscriptionService } from './voice-transcription.service';
 
 @Injectable()
 export class SchedulesService {
@@ -33,6 +31,7 @@ export class SchedulesService {
     private categoryRepository: Repository<Category>,
     private readonly httpService: HttpService,
     private readonly configService: ConfigService,
+    private readonly voiceTranscriptionService: VoiceTranscriptionService,
   ) {
     const openaiApiKey = this.configService.get<string>('OPENAI_API_KEY');
     this.openai = new OpenAI({
@@ -226,142 +225,6 @@ export class SchedulesService {
     return dtos;
   }
 
-  async transcribeAudio(file: Express.Multer.File): Promise<string> {
-    const jwtToken = await this.ensureValidToken();
-
-    const formData = new FormData();
-    formData.append('file', file.buffer, {
-      filename: file.originalname,
-      contentType: file.mimetype,
-    });
-    formData.append(
-      'config',
-      JSON.stringify({
-        use_diarization: true,
-        use_itn: true,
-        use_disfluency_filter: false,
-        use_profanity_filter: false,
-        use_paragraph_splitter: true,
-        paragraph_splitter: { max: 50 },
-        keywords: ['내일', '모레', '글피', '내년', '올해', '오늘'],
-      }),
-    );
-
-    const url = 'https://openapi.vito.ai/v1/transcribe';
-
-    try {
-      const response = await lastValueFrom(
-        this.httpService.post(url, formData, {
-          headers: {
-            Authorization: `Bearer ${jwtToken}`,
-            ...formData.getHeaders(),
-          },
-        }),
-      );
-      return response.data.id;
-    } catch (error) {
-      if (error.response?.status === 401) {
-        // 토큰이 만료되었을 경우, 토큰을 갱신하고 다시 시도
-        console.log('Token expired, refreshing and retrying...');
-        await this.ensureValidToken(); // 강제로 새 토큰 발급
-        return this.transcribeAudio(file); // 재귀적으로 다시 시도
-      }
-      throw new Error(`STT 요청 실패: ${error.message}`);
-    }
-  }
-
-  // 전사 결과 조회
-  async getTranscriptionResult(transcribeId: string): Promise<any> {
-    const jwtToken = await this.ensureValidToken();
-
-    const url = `https://openapi.vito.ai/v1/transcribe/${transcribeId}`;
-
-    try {
-      const response = await lastValueFrom(
-        this.httpService.get(url, {
-          headers: {
-            Authorization: `Bearer ${jwtToken}`,
-          },
-        }),
-      );
-      return response.data;
-    } catch (error) {
-      if (error.response?.status === 401) {
-        // 토큰이 만료되었을 경우, 토큰을 갱신하고 다시 시도
-        console.log('Token expired, refreshing and retrying...');
-        await this.ensureValidToken(); // 강제로 새 토큰 발급
-        return this.getTranscriptionResult(transcribeId); // 재귀적으로 다시 시도
-      }
-      throw new Error(`전사 결과 조회 실패: ${error.message}`);
-    }
-  }
-
-  private isTokenExpiredOrCloseToExpiry(
-    token: string,
-    thresholdMinutes: number = 5,
-  ): boolean {
-    try {
-      const decoded = jwt.decode(token) as { exp: number };
-      if (!decoded || !decoded.exp) {
-        return true; // 토큰이 유효하지 않거나 만료 시간이 없으면 만료된 것으로 간주
-      }
-
-      const currentTime = Math.floor(Date.now() / 1000);
-      const timeUntilExpiry = decoded.exp - currentTime;
-      const thresholdSeconds = thresholdMinutes * 60;
-
-      return timeUntilExpiry <= thresholdSeconds;
-    } catch (error) {
-      console.error('토큰 검증 중 오류 발생:', error);
-      return true; // 오류 발생 시 만료된 것으로 간주
-    }
-  }
-
-  async ensureValidToken(): Promise<string> {
-    // 토큰 유효성 검사 로직
-    // 만료되었거나 곧 만료될 예정이면 새로운 토큰 발급
-    const currentToken = this.configService.get<string>(
-      'RETURN_ZERO_ACCESS_TOKEN',
-    );
-    // 토큰 만료 시간 확인 로직 추가 필요
-    if (this.isTokenExpiredOrCloseToExpiry(currentToken)) {
-      const newToken = await this.getJwtToken();
-      // 새 토큰을 환경 변수나 설정에 저장
-      this.configService.set('RETURN_ZERO_ACCESS_TOKEN', newToken);
-      return newToken;
-    }
-    return currentToken;
-  }
-
-  // 전사 결과가 완료될 때까지 폴링
-  async pollTranscriptionResult(
-    transcribeId: string,
-    maxRetries: number = 10,
-    interval: number = 20000,
-  ): Promise<any> {
-    for (let attempt = 0; attempt < maxRetries; attempt++) {
-      try {
-        const result = await this.getTranscriptionResult(transcribeId);
-        if (result.status === 'completed') {
-          return result;
-        }
-      } catch (error) {
-        console.error(`시도 ${attempt + 1} failed:`, error);
-        if (attempt === maxRetries - 1) {
-          throw error;
-        }
-      }
-      await new Promise((resolve) => setTimeout(resolve, interval));
-    }
-    throw new Error('전사 결과를 완료할 수 없습니다. 시간이 초과되었습니다.');
-  }
-
-  // 파일 업로드 후 전사 및 결과 조회
-  async transcribeAndFetchResult(file: Express.Multer.File): Promise<any> {
-    const transcribeId = await this.transcribeAudio(file);
-    return await this.pollTranscriptionResult(transcribeId);
-  }
-
   private parseGptResponse(response: string): any[] {
     try {
       // GPT 응답을 JSON으로 파싱
@@ -369,32 +232,6 @@ export class SchedulesService {
     } catch (error) {
       console.error('Error parsing GPT response:', error);
       throw new Error('Failed to parse GPT response');
-    }
-  }
-
-  async getJwtToken(): Promise<string> {
-    const clientId = this.configService.get<string>('RETURN_ZERO_CLIENT_ID');
-    const clientSecret = this.configService.get<string>(
-      'RETURN_ZERO_CLIENT_SECRET',
-    );
-
-    const url = 'https://openapi.vito.ai/v1/authenticate'; // RTZR 인증 URL
-
-    try {
-      const response = await lastValueFrom(
-        this.httpService.post(
-          url,
-          `client_id=${clientId}&client_secret=${clientSecret}`,
-          {
-            headers: {
-              'Content-Type': 'application/x-www-form-urlencoded',
-            },
-          },
-        ),
-      );
-      return response.data.access_token; // JWT 토큰 반환
-    } catch (error) {
-      throw new Error(`JWT 토큰 발급 실패: ${error.message}`);
     }
   }
 
@@ -407,10 +244,9 @@ export class SchedulesService {
       apiKey: this.configService.get<string>('OPENAI_API_KEY'),
     });
 
-    const transcriptionText = transcriptionResult.results.utterances
-      .map((u) => u.msg)
-      .join(' ');
-    const formattedDate = formatDateToYYYYMMDDHHMMSS(new Date(currentDateTime));
+    const formattedDate = this.formatDateToYYYYMMDDHHMMSS(
+      new Date(currentDateTime),
+    );
 
     const gptResponse = await openai.chat.completions.create({
       model: this.configService.get<string>('OPENAI_FINETUNING_MODEL'),
@@ -422,7 +258,7 @@ export class SchedulesService {
         },
         {
           role: 'user',
-          content: `{Today : ${formattedDate}, conversations : ${transcriptionText}}`,
+          content: `{Today : ${formattedDate}, conversations : ${transcriptionResult}}`,
         },
       ],
     });
@@ -431,6 +267,20 @@ export class SchedulesService {
     return this.convertGptResponseToCreateDto(
       this.parseGptResponse(gptResponseContent),
     );
+  }
+
+  // 날짜를 YYYY-MM-DD HH:mm:ss 형식으로 변환하는 함수
+  private async formatDateToYYYYMMDDHHMMSS(date: Date): Promise<string> {
+    const year = date.getFullYear();
+    const month = String(date.getMonth() + 1).padStart(2, '0'); // 월은 0부터 시작하므로 +1
+    const day = String(date.getDate()).padStart(2, '0');
+    const hours = String(date.getHours()).padStart(2, '0');
+    const minutes = String(date.getMinutes()).padStart(2, '0');
+    const seconds = String(date.getSeconds()).padStart(2, '0');
+    // console.log(
+    //   `date지렁 : ${year}-${month}-${day} ${hours}:${minutes}:${seconds}`,
+    // );
+    return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
   }
 
   private async convertGptResponseToCreateDto(
@@ -458,31 +308,27 @@ export class SchedulesService {
         // 카테고리 매핑
         const category = categoryMap[event.category] || categoryMap['기타'];
         dto.category = category; // 전체 카테고리 객체 설정
-
-        console.log(dto);
         return dto;
       }),
     );
   }
-  async transcribeAndFetchResultWithGpt(
+  async transcribeRTZRAndFetchResultWithGpt(
     file: Express.Multer.File,
     currentDateTime: string,
   ) {
-    const transcribe = await this.transcribeAndFetchResult(file);
+    const transcribe =
+      await this.voiceTranscriptionService.RTZRTranscribeResult(file);
     const result = await this.processWithGpt(transcribe, currentDateTime);
     return result;
   }
-}
-// 날짜를 YYYY-MM-DD HH:mm:ss 형식으로 변환하는 함수
-function formatDateToYYYYMMDDHHMMSS(date: Date): string {
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, '0'); // 월은 0부터 시작하므로 +1
-  const day = String(date.getDate()).padStart(2, '0');
-  const hours = String(date.getHours()).padStart(2, '0');
-  const minutes = String(date.getMinutes()).padStart(2, '0');
-  const seconds = String(date.getSeconds()).padStart(2, '0');
-  // console.log(
-  //   `date지렁 : ${year}-${month}-${day} ${hours}:${minutes}:${seconds}`,
-  // );
-  return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
+
+  async transcribeWhisperAndFetchResultWithGpt(
+    file: Express.Multer.File,
+    currentDateTime: string,
+  ) {
+    const transcribe =
+      await this.voiceTranscriptionService.whisperTranscribeResult(file);
+    const result = await this.processWithGpt(transcribe, currentDateTime);
+    return result;
+  }
 }
